scripts: setup-apt: fix non-free component and stop handling contrib
The `non-free` component is now `non-free-firmware` and should be
added to the Mobian sources file whenever needed. It shouldn't be part
of the `debian-security` entry though, as it only supports `main`.

As Mobian doesn't provide a `contrib` component and this one isn't
used for `debian-security` for the reason stated above, we can also stop
bothering with it in this script and only handle `non-free-firmware`.