scripts: split out non-free-firmware setup from setup-apt.sh
The current way we handle the `non-free-firmware` component is buggy as
the Mobian part of it can only be setup in the initial `rootfs` recipe;
enabling it afterwards will only affect the Debian repo setup. This used
to work as expected as `non-free-firmware` was always enabled for Mobian
even though it wasn't always needed.

This creates a new `setup-apt-nonfree.sh` which handles setting up this
component for both the Debian and Mobian repos, ensuring they stay in
sync.