devices: librem5: bootloader: stop installing u-boot
The current version of `librem5-flash-image` is able to install `u-boot`
to the eMMC boot partition, so we can just recommend using it for
installing Mobian and stop worrying about that part.

This also means we can simply go with the same process we use for PINE64
devices, as in "just ensure extlinux.conf is valid".

Also drop `bootstart` image build option as it's no longer needed when
not flashing the bootloader to the image.