devices: qcom: configs: generate more boot images
We only build boot images for a subset of the devices supported by our
kernels. Let's ensure we actually build (and publish) all possible boot
images for a given kernel so it makes it easier for people to test their
devices with Mobian.

Note: `spacewar` is part of the `sc7280-mainline` tree but not present
in our patches, so keep it commented out for now.